+++ microbench/src/main/java/org/apache/pulsar/broker/qos/AsyncTokenBucketBenchmark.java
@@ -20,6 +20,9 @@
 package org.apache.pulsar.broker.qos;
 
 import java.util.concurrent.TimeUnit;
+
+import org.apache.pulsar.common.util.qos.AsyncTokenBucket;
+import org.apache.pulsar.common.util.qos.DefaultMonotonicSnapshotClock;
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.BenchmarkMode;
 import org.openjdk.jmh.annotations.Fork;
