+++ pulsar-broker/src/main/java/org/apache/pulsar/broker/stats/WindowWrap.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.common.util;
+package org.apache.pulsar.broker.stats;
 
 public final class WindowWrap<T> {
     private final long interval;
