+++ pulsar-common/src/main/java/org/apache/pulsar/common/util/qos/DynamicRateAsyncTokenBucket.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.pulsar.broker.qos;
+package org.apache.pulsar.common.util.qos;
 
 import java.util.function.LongSupplier;
