+++ pulsar-common/src/test/java/org/apache/pulsar/common/util/qos/AsyncTokenBucketTest.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.pulsar.broker.qos;
+package org.apache.pulsar.common.util.qos;
 
 import static org.testng.Assert.assertEquals;
 import java.util.concurrent.TimeUnit;
